+++ dyninstAPI/CMakeLists.txt
@@ -268,6 +268,8 @@ if(DYNINST_CODEGEN_ARCH_I386 OR DYNINST_CODEGEN_ARCH_X86_64)
   list(
     APPEND
     _sources
+    src/codegen/emitters/x86/IA32/EmitterIA32.C
+    src/codegen/emitters/x86/IA32/EmitterIA32.h
     src/codegen/emitters/x86/Emitterx86.C
     src/codegen/emitters/x86/Emitterx86.h
     src/codegen/emitters/x86/generators.C
